Wait for a complete timer delegate thread stop
In some cases when timer just stops delegate thread without waiting
for it's complete finishing, there might be some unexpected data races
which can be avoided if thread is finished completely. By that reason,
thread stopping priority should be increased to soft stop.
As it causes some random fails in Linux build, making this change an
Android specific for now.